Fix #1463: add apis and client as submodule
