+++ pkg/apis/camel/go.mod
@@ -1,0 +1,8 @@
+module github.com/apache/camel-k/pkg/apis/camel
+
+go 1.13
+
+require (
+	k8s.io/api v0.16.4
+	k8s.io/apimachinery v0.16.4
+)
