+++ pkg/apis/camel/v1/build_types.go
@@ -171,7 +171,3 @@ type BuildCondition struct {
 	// A human readable message indicating details about the transition.
 	Message string `json:"message,omitempty"`
 }
-
-func init() {
-	SchemeBuilder.Register(&Build{}, &BuildList{})
-}
