+++ pkg/apis/camel/v1/integration_types.go
@@ -260,7 +260,3 @@ type IntegrationCondition struct {
 	// A human readable message indicating details about the transition.
 	Message string `json:"message,omitempty"`
 }
-
-func init() {
-	SchemeBuilder.Register(&Integration{}, &IntegrationList{})
-}
