+++ pkg/apis/camel/v1/integrationplatform_types.go
@@ -202,7 +202,3 @@ type IntegrationPlatformCondition struct {
 	// A human readable message indicating details about the transition.
 	Message string `json:"message,omitempty"`
 }
-
-func init() {
-	SchemeBuilder.Register(&IntegrationPlatform{}, &IntegrationPlatformList{})
-}
