+++ pkg/apis/camel/v1/register.go
@@ -23,8 +23,9 @@ limitations under the License.
 package v1
 
 import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"sigs.k8s.io/controller-runtime/pkg/scheme"
 )
 
 var (
@@ -32,7 +33,7 @@ var (
 	SchemeGroupVersion = schema.GroupVersion{Group: "camel.apache.org", Version: "v1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
-	SchemeBuilder = &scheme.Builder{GroupVersion: SchemeGroupVersion}
+	SchemeBuilder = runtime.NewSchemeBuilder(addKnownTypes)
 
 	// AddToScheme is a shortcut to SchemeBuilder.AddToScheme
 	AddToScheme = SchemeBuilder.AddToScheme
@@ -41,4 +42,22 @@ var (
 // Resource takes an unqualified resource and returns a Group qualified GroupResource
 func Resource(resource string) schema.GroupResource {
 	return SchemeGroupVersion.WithResource(resource).GroupResource()
+}
+
+// Adds the list of known types to Scheme.
+func addKnownTypes(scheme *runtime.Scheme) error {
+	scheme.AddKnownTypes(SchemeGroupVersion,
+		&Integration{},
+		&IntegrationList{},
+		&IntegrationKit{},
+		&IntegrationKitList{},
+		&IntegrationPlatform{},
+		&IntegrationPlatformList{},
+		&CamelCatalog{},
+		&CamelCatalogList{},
+		&Build{},
+		&BuildList{},
+	)
+	metav1.AddToGroupVersion(scheme, SchemeGroupVersion)
+	return nil
 }
