+++ pkg/client/camel/clientset/versioned/doc.go
