+++ pkg/client/camel/clientset/versioned/fake/doc.go
