+++ pkg/client/camel/clientset/versioned/fake/register.go
