+++ pkg/client/camel/clientset/versioned/scheme/doc.go
