+++ pkg/client/camel/clientset/versioned/scheme/register.go
