+++ pkg/client/camel/clientset/versioned/typed/camel/v1/doc.go
