+++ pkg/client/camel/clientset/versioned/typed/camel/v1/fake/doc.go
