+++ pkg/client/camel/clientset/versioned/typed/camel/v1/fake/fake_integrationkit.go
