+++ pkg/client/camel/clientset/versioned/typed/camel/v1/generated_expansion.go
