+++ pkg/client/camel/informers/externalversions/generic.go
