+++ pkg/client/camel/informers/externalversions/internalinterfaces/factory_interfaces.go
@@ -22,7 +22,7 @@ package internalinterfaces
 import (
 	time "time"
 
-	versioned "github.com/apache/camel-k/pkg/client/clientset/versioned"
+	versioned "github.com/apache/camel-k/pkg/client/camel/clientset/versioned"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	cache "k8s.io/client-go/tools/cache"
