+++ pkg/client/camel/listers/camel/v1/expansion_generated.go
