+++ pkg/client/camel/listers/camel/v1/integration.go
