+++ pkg/client/camel/listers/camel/v1/integrationkit.go
