+++ pkg/client/camel/listers/camel/v1/integrationplatform.go
